Fix warning

+++ LMAlertViewDemo/LMModalViewController.m
@@ -53,8 +53,8 @@
 
 - (void)textViewDidChange:(UITextView *)textView
 {
-	int remainingCharacters = 140 - textView.text.length;
-	self.characterCountLabel.text = [NSString stringWithFormat:@"%i", remainingCharacters];
+	NSUInteger remainingCharacters = 140 - textView.text.length;
+	self.characterCountLabel.text = [NSString stringWithFormat:@"%lu", (unsigned long)remainingCharacters];
 	
 	self.navigationItem.rightBarButtonItem.enabled = (remainingCharacters >= 0);
 }
